@@ -2,29 +2,28 @@
 
 namespace App\Filament\App\Resources\UserResource\Pages;
 
+use App\Filament\App\Resources\UserResource;
 use App\Models\User;
 use Filament\Actions;
-use Filament\Support\Enums\Alignment;
 use Filament\Notifications\Notification;
 use Filament\Resources\Pages\EditRecord;
-use App\Filament\App\Resources\UserResource;
+use Filament\Support\Enums\Alignment;
 
 class EditUser extends EditRecord
 {
     protected static string $resource = UserResource::class;
 
-    public static string | Alignment $formActionsAlignment = Alignment::Right;
+    public static string|Alignment $formActionsAlignment = Alignment::Right;
 
     protected function getHeaderActions(): array
     {
         return [
             Actions\DeleteAction::make()
-                ->successNotification(fn ($record) =>
-                    Notification::make()
-                        ->success()
-                        ->title($record->name . ' deleted')
-                        ->body('The user has been deleted successfully.')
-                        ->sendToDatabase(User::role('admin')->get()),
+                ->successNotification(fn ($record) => Notification::make()
+                    ->success()
+                    ->title($record->name.' deleted')
+                    ->body('The user has been deleted successfully.')
+                    ->sendToDatabase(User::role('admin')->get()),
                 ),
         ];
     }
@@ -38,7 +37,7 @@ class EditUser extends EditRecord
     {
         return Notification::make()
             ->info()
-            ->title($this->getRecord()->name . ' updated')
+            ->title($this->getRecord()->name.' updated')
             ->body('Changes have been saved successfully.')
             ->sendToDatabase(User::role('admin')->get());
     }
