@@ -5,11 +5,10 @@ namespace App\Mail;
 use App\Models\User;
 use Illuminate\Bus\Queueable;
 use Illuminate\Mail\Mailable;
-use Illuminate\Support\Facades\URL;
 use Illuminate\Mail\Mailables\Content;
-use Illuminate\Queue\SerializesModels;
 use Illuminate\Mail\Mailables\Envelope;
-use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Facades\URL;
 
 class TeamInvitationMail extends Mailable
 {
@@ -29,23 +28,23 @@ class TeamInvitationMail extends Mailable
      * Get the message envelope.
      */
     public function envelope(): Envelope
-{
-    return new Envelope(
-        subject: 'Invitation to join ' . config('app.name'),
-    );
-}
+    {
+        return new Envelope(
+            subject: 'Invitation to join '.config('app.name'),
+        );
+    }
 
-public function content(): Content
-{
-    return new Content(
-        view: 'view.name',
-        markdown: 'emails.team-invitation',
-        with: [
-            'acceptUrl' => URL::signedRoute(
-                "invitation.accept",
-                ['invitation' => $this->user]
-            ),
-        ]
-    );
-}
+    public function content(): Content
+    {
+        return new Content(
+            view: 'view.name',
+            markdown: 'emails.team-invitation',
+            with: [
+                'acceptUrl' => URL::signedRoute(
+                    'invitation.accept',
+                    ['invitation' => $this->user]
+                ),
+            ]
+        );
+    }
 }
