@@ -3,8 +3,8 @@
 namespace App\Policies;
 
 use App\Models\User;
-use Spatie\Permission\Models\Role;
 use Illuminate\Auth\Access\HandlesAuthorization;
+use Spatie\Permission\Models\Role;
 
 class RolePolicy
 {
@@ -12,9 +12,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can view any models.
-     *
-     * @param  \App\Models\User  $user
-     * @return bool
      */
     public function viewAny(User $user): bool
     {
@@ -23,10 +20,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can view the model.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \Spatie\Permission\Models\Role  $role
-     * @return bool
      */
     public function view(User $user, Role $role): bool
     {
@@ -35,9 +28,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can create models.
-     *
-     * @param  \App\Models\User  $user
-     * @return bool
      */
     public function create(User $user): bool
     {
@@ -46,10 +36,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can update the model.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \Spatie\Permission\Models\Role  $role
-     * @return bool
      */
     public function update(User $user, Role $role): bool
     {
@@ -58,10 +44,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can delete the model.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \Spatie\Permission\Models\Role  $role
-     * @return bool
      */
     public function delete(User $user, Role $role): bool
     {
@@ -70,9 +52,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can bulk delete.
-     *
-     * @param  \App\Models\User  $user
-     * @return bool
      */
     public function deleteAny(User $user): bool
     {
@@ -81,10 +60,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can permanently delete.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \Spatie\Permission\Models\Role  $role
-     * @return bool
      */
     public function forceDelete(User $user, Role $role): bool
     {
@@ -93,9 +68,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can permanently bulk delete.
-     *
-     * @param  \App\Models\User  $user
-     * @return bool
      */
     public function forceDeleteAny(User $user): bool
     {
@@ -104,10 +76,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can restore.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \Spatie\Permission\Models\Role  $role
-     * @return bool
      */
     public function restore(User $user, Role $role): bool
     {
@@ -116,9 +84,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can bulk restore.
-     *
-     * @param  \App\Models\User  $user
-     * @return bool
      */
     public function restoreAny(User $user): bool
     {
@@ -127,10 +92,6 @@ class RolePolicy
 
     /**
      * Determine whether the user can replicate.
-     *
-     * @param  \App\Models\User  $user
-     * @param  \Spatie\Permission\Models\Role  $role
-     * @return bool
      */
     public function replicate(User $user, Role $role): bool
     {
@@ -139,13 +100,9 @@ class RolePolicy
 
     /**
      * Determine whether the user can reorder.
-     *
-     * @param  \App\Models\User  $user
-     * @return bool
      */
     public function reorder(User $user): bool
     {
         return $user->can('{{ Reorder }}');
     }
-
 }
