@@ -2,16 +2,15 @@
 
 namespace Database\Seeders;
 
-use Illuminate\Database\Seeder;
 use BezhanSalleh\FilamentShield\Support\Utils;
+use Illuminate\Database\Seeder;
+
 class ShieldSeeder extends Seeder
 {
     /**
-    * Run the database seeds.
-    *
-    * @return void
-    */
-    public function run()
+     * Run the database seeds.
+     */
+    public function run(): void
     {
         app()[\Spatie\Permission\PermissionRegistrar::class]->forgetCachedPermissions();
 
@@ -26,12 +25,12 @@ class ShieldSeeder extends Seeder
 
     protected static function makeRolesWithPermissions(string $rolesWithPermissions): void
     {
-        if (! blank($rolePlusPermissions = json_decode($rolesWithPermissions,true))) {
+        if (! blank($rolePlusPermissions = json_decode($rolesWithPermissions, true))) {
 
             foreach ($rolePlusPermissions as $rolePlusPermission) {
                 $role = Utils::getRoleModel()::firstOrCreate([
                     'name' => $rolePlusPermission['name'],
-                    'guard_name' => $rolePlusPermission['guard_name']
+                    'guard_name' => $rolePlusPermission['guard_name'],
                 ]);
 
                 if (! blank($rolePlusPermission['permissions'])) {
@@ -39,10 +38,10 @@ class ShieldSeeder extends Seeder
                     $permissionModels = collect();
 
                     collect($rolePlusPermission['permissions'])
-                        ->each(function ($permission) use($permissionModels) {
+                        ->each(function ($permission) use ($permissionModels) {
                             $permissionModels->push(Utils::getPermissionModel()::firstOrCreate([
                                 'name' => $permission,
-                                'guard_name' => 'web'
+                                'guard_name' => 'web',
                             ]));
                         });
                     $role->syncPermissions($permissionModels);
@@ -54,9 +53,9 @@ class ShieldSeeder extends Seeder
 
     public static function makeDirectPermissions(string $directPermissions): void
     {
-        if (! blank($permissions = json_decode($directPermissions,true))) {
+        if (! blank($permissions = json_decode($directPermissions, true))) {
 
-            foreach($permissions as $permission) {
+            foreach ($permissions as $permission) {
 
                 if (Utils::getPermissionModel()::whereName($permission)->doesntExist()) {
                     Utils::getPermissionModel()::create([
